feedback addressed
