+++ sources/packages/api/src/utilities/index.ts
@@ -3,4 +3,4 @@ export * from "./string-builder";
 export * from "./custom-named-error";
 export * from "./date-utils";
 export * from "./system-configurations-constants";
-export * from "./application";
+export * from "./application-utils";
